% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noRdic.R
\name{read.nordic}
\alias{read.nordic}
\title{read.nordic}
\usage{
read.nordic(filename, ...)
}
\arguments{
\item{filename}{- filename to load}

\item{...}{- other arguments passed to read.csv}
}
\value{
a dataframe
}
\description{
Reads a Nordic data CSV in this format:
}
\details{
AREA GENUSE       SPUSE
  1  SWE      F F_alcoholic
  2  ICE      F F_alcoholic

... and converts it to a long format table:

     AREA   TYPE                USE
 1   SWE GENUSE                 F
 2   ICE GENUSE                 F
}
