% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noRdic.R
\name{to.rayDISC}
\alias{to.rayDISC}
\title{to.rayDISC}
\usage{
to.rayDISC(df, taxa, state = "0")
}
\arguments{
\item{df}{- a dataframe}

\item{taxa}{- a vector of taxa that should be present.}

\item{state}{- the state code to add (default= "0")}
}
\value{
a dataframe
}
\description{
Constructs a data frame for rayDISC, adding in missing taxa
}
\details{
AREA   TYPE          USE
 1  NORN GENUSE           IC
 2   SWE GENUSE           IC
 3   DEN GENUSE           IC

... and this command:

  > augment(df, c("NORN", "SWE", "DEN", "A", "B"))

    Taxon State
    AREA   TYPE          USE    STATE
 1  NORN GENUSE           IC       1
 2   SWE GENUSE           IC       1
 3   DEN GENUSE           IC       1
 4     A GENUSE           IC       0
 5     B GENUSE           IC       0
}
